Add files via upload
My circuit realizes that when the switch is turned on, the LED lights up to remind that as time goes on, the frequency of LED lights will become faster and faster. It is very similar to the reversing radar. The closer the car is to an obstacle, the louder the frequency of the radar warning.
In the future, I think there will be educational robots to replace the work of some teachers, and robots will accompany children to learn some social rules, or welding circuits and other technologies. When he makes a mistake, the indicator lights up to remind the children to correct their behavior.